Move MapType to model package
