+++ app/src/main/java/org/groundplatform/android/persistence/local/LocalValueStore.kt
@@ -25,7 +25,7 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import org.groundplatform.android.model.map.CameraPosition
-import org.groundplatform.android.ui.map.MapType
+import org.groundplatform.android.model.map.MapType
 import org.groundplatform.android.ui.settings.Keys
 import org.groundplatform.android.util.allowThreadDiskReads
 import org.groundplatform.android.util.allowThreadDiskWrites
