+++ app/src/main/java/org/groundplatform/android/repository/MapStateRepository.kt
@@ -19,8 +19,8 @@ import javax.inject.Inject
 import javax.inject.Singleton
 import kotlinx.coroutines.flow.Flow
 import org.groundplatform.android.model.map.CameraPosition
+import org.groundplatform.android.model.map.MapType
 import org.groundplatform.android.persistence.local.LocalValueStore
-import org.groundplatform.android.ui.map.MapType
 
 /** Provides access and storage of persistent map states. */
 @Singleton
