+++ app/src/main/java/org/groundplatform/android/ui/home/mapcontainer/MapTypeDialogFragment.kt
@@ -24,8 +24,8 @@ import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 import org.groundplatform.android.databinding.MapTypeDialogFragmentBinding
+import org.groundplatform.android.model.map.MapType
 import org.groundplatform.android.ui.common.ViewModelFactory
-import org.groundplatform.android.ui.map.MapType
 
 /** Dialog fragment containing a list of [MapType] for updating basemap layer. */
 @AndroidEntryPoint
