+++ app/src/main/java/org/groundplatform/android/ui/map/MapFragment.kt
@@ -23,6 +23,7 @@ import org.groundplatform.android.model.imagery.TileSource
 import org.groundplatform.android.model.map.Bounds
 import org.groundplatform.android.model.map.CameraPosition
 import org.groundplatform.android.model.map.Feature
+import org.groundplatform.android.model.map.MapType
 import org.groundplatform.android.ui.common.AbstractFragment
 
 /** Implementation of Fragment which supports displaying a map. */
