+++ app/src/main/java/org/groundplatform/android/ui/offlineareas/selector/OfflineAreaSelectorFragment.kt
@@ -31,13 +31,13 @@ import javax.inject.Inject
 import kotlinx.coroutines.launch
 import org.groundplatform.android.R
 import org.groundplatform.android.databinding.OfflineAreaSelectorFragBinding
+import org.groundplatform.android.model.map.MapType
 import org.groundplatform.android.ui.common.AbstractMapContainerFragment
 import org.groundplatform.android.ui.common.BaseMapViewModel
 import org.groundplatform.android.ui.common.EphemeralPopups
 import org.groundplatform.android.ui.common.MapConfig
 import org.groundplatform.android.ui.home.mapcontainer.HomeScreenMapContainerViewModel
 import org.groundplatform.android.ui.map.MapFragment
-import org.groundplatform.android.ui.map.MapType
 import org.groundplatform.android.util.renderComposableDialog
 
 /** Map UI used to select areas for download and viewing offline. */
