+++ app/src/test/java/org/groundplatform/android/repository/MapStateRepositoryTest.kt
@@ -22,8 +22,8 @@ import javax.inject.Inject
 import org.groundplatform.android.BaseHiltTest
 import org.groundplatform.android.FakeData
 import org.groundplatform.android.model.map.CameraPosition
+import org.groundplatform.android.model.map.MapType
 import org.groundplatform.android.persistence.local.LocalValueStore
-import org.groundplatform.android.ui.map.MapType
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
