+++ app/src/test/java/org/groundplatform/android/ui/offlineareas/viewer/OfflineAreaViewerFragmentTest.kt
@@ -33,9 +33,9 @@ import org.groundplatform.android.BaseHiltTest
 import org.groundplatform.android.FakeData.OFFLINE_AREA
 import org.groundplatform.android.R
 import org.groundplatform.android.launchFragmentWithNavController
+import org.groundplatform.android.model.map.MapType
 import org.groundplatform.android.persistence.local.stores.LocalOfflineAreaStore
 import org.groundplatform.android.ui.common.MapConfig
-import org.groundplatform.android.ui.map.MapType
 import org.groundplatform.android.util.view.isGone
 import org.junit.Test
 import org.junit.runner.RunWith
